Correção do tipo da variável.

+++ Desafio31.py
@@ -1,4 +1,4 @@
-distancia = int(input("Digite a distância em Kilometros que você vai fazer: "))
+distancia = float(input("Digite a distância em Kilometros que você vai fazer: "))
 if distancia <= 200:
     distancia = distancia * 0.50
     print(f"Sua passagem ficou no valor de R${distancia}!")
